fix: adding injectable

+++ server/src/application/use-cases/address/delete-address/delete-address.use-case.ts
@@ -1,5 +1,7 @@
 import { AddressRepository } from '@application/repositories/address/address.repository';
 import { FindAddressUseCase } from '../find-address/find-address.use-case';
+import { Injectable } from '@nestjs/common';
+import { ObjectNotFoundException } from '@application/exceptions/object-not-found.exception';
 
 interface DeleteAddressUseCaseRequest {
   id: string;
@@ -7,16 +9,20 @@ interface DeleteAddressUseCaseRequest {
 
 type DeleteAddressUseCaseResponse = void;
 
+@Injectable()
 export class DeleteAddressUseCase {
-  constructor(
-    private addressRepository: AddressRepository,
-    private findAddressUseCase: FindAddressUseCase,
-  ) {}
+  constructor(private addressRepository: AddressRepository) {}
 
   async execute(
     request: DeleteAddressUseCaseRequest,
   ): Promise<DeleteAddressUseCaseResponse> {
-    const { address } = await this.findAddressUseCase.execute(request);
+    const { id } = request;
+    const address = await this.addressRepository.findById(id);
+
+    if (!address) {
+      throw new ObjectNotFoundException('Address was not found.');
+    }
+
     await this.addressRepository.delete(address.id);
   }
 }
